Refactor grpcclient to gmsacredclient
