+++ agent/taskresource/credentialspec/credentialspec_linux.go
@@ -37,7 +37,7 @@ import (
 	s3factory "github.com/aws/amazon-ecs-agent/agent/s3/factory"
 	ssmfactory "github.com/aws/amazon-ecs-agent/agent/ssm/factory"
 	"github.com/aws/amazon-ecs-agent/ecs-agent/credentials"
-	credentialsfetcherclient "github.com/aws/amazon-ecs-agent/ecs-agent/grpcclient"
+	credentialsfetcherclient "github.com/aws/amazon-ecs-agent/ecs-agent/gmsacredclient"
 	"github.com/pkg/errors"
 )
