+++ ecs-agent/gmsacredclient/credentialsfetcher/credentialsfetcher.proto
@@ -4,7 +4,7 @@
 // This will generate credentialsfetcher/credentialsfetcher.pb.go and credentialsfetcher/credentialsfetcher_grpc.pb.go files
 syntax = "proto3";
 
-option go_package = "grpcclient/credentialsfetcher";
+option go_package = "gmsacredclient/credentialsfetcher";
 package credentialsfetcher;
 
 service CredentialsFetcherService {
