+++ ecs-agent/gmsacredclient/credentialsfetcherclient_test.go
@@ -13,7 +13,7 @@
 // on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 // express or implied. See the License for the specific language governing
 // permissions and limitations under the License.
-package grpcclient
+package gmsacredclient
 
 import (
 	"context"
@@ -22,7 +22,7 @@ import (
 	"testing"
 	"time"
 
-	pb "github.com/aws/amazon-ecs-agent/ecs-agent/grpcclient/credentialsfetcher"
+	pb "github.com/aws/amazon-ecs-agent/ecs-agent/gmsacredclient/credentialsfetcher"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc"
